Organize, superficial refactor
