+++ app/src/main/java/com/copixelate/ui/screens/contacts/ContactsScreen.kt
@@ -20,8 +20,8 @@ import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
+import com.copixelate.ui.common.AddItemFab
 import com.copixelate.ui.nav.NavInfo
-import com.copixelate.ui.screens.AddItemFab
 import com.copixelate.ui.util.PreviewSurface
 import com.copixelate.ui.util.toDp
