+++ app/src/main/java/com/copixelate/ui/screens/library/CreateItemDialog.kt
@@ -1,250 +1,50 @@
-package com.copixelate.ui.screens
+package com.copixelate.ui.screens.library
 
-import androidx.compose.animation.core.LinearOutSlowInEasing
-import androidx.compose.animation.core.animateFloatAsState
-import androidx.compose.animation.core.tween
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.itemsIndexed
-import androidx.compose.foundation.lazy.rememberLazyListState
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.defaultMinSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.foundation.layout.wrapContentWidth
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.*
+import androidx.compose.material.icons.filled.KeyboardArrowLeft
+import androidx.compose.material.icons.filled.KeyboardArrowRight
+import androidx.compose.material.icons.filled.KeyboardDoubleArrowLeft
+import androidx.compose.material.icons.filled.KeyboardDoubleArrowRight
 import androidx.compose.material.icons.outlined.AddPhotoAlternate
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.AlertDialogDefaults
+import androidx.compose.material3.Button
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Slider
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.alpha
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.setValue
-import androidx.navigation.NavHostController
-import com.copixelate.art.ArtSpace
-import com.copixelate.data.model.SpaceModel
-import com.copixelate.data.model.toArtSpace
-import com.copixelate.data.model.toModel
-import com.copixelate.ui.components.BitmapImage
-import com.copixelate.ui.nav.NavInfo
-import com.copixelate.ui.nav.navigate
 import com.copixelate.ui.util.PreviewSurface
-import com.copixelate.ui.util.toDp
-import com.copixelate.viewmodel.LibraryViewModel
-
-
-@Composable
-fun LibraryScreen(navController: NavHostController, libraryViewModel: LibraryViewModel) {
-    LibraryScreenContent(
-        spaces = libraryViewModel.allSpaces.collectAsState().value,
-        onCreate = { width, height, paletteSize ->
-            libraryViewModel.createNewArtSpace(width, height, paletteSize)
-        },
-        onDelete = { spaceModel ->
-            libraryViewModel.deleteArtSpace(spaceModel)
-        },
-        onOpen = { spaceModel ->
-            libraryViewModel.updateCurrentSpaceId(spaceModel.id)
-            navController.navigate(navInfo = NavInfo.Art)
-        }
-    )
-}
-
-@Composable
-private fun LibraryScreenContent(
-    spaces: List<SpaceModel>,
-    onCreate: (width: Int, height: Int, paletteSize: Int) -> Unit,
-    onDelete: (SpaceModel) -> Unit,
-    onOpen: (SpaceModel) -> Unit,
-) {
-
-    val scrollState = rememberLazyListState()
-    var addItemJustOccurred by remember { mutableStateOf(false) }
-
-    var fabHeight by remember { mutableStateOf(0) }
-    val fabClearance = fabHeight.toDp() + 16.dp + 16.dp
-
-    // Scroll to the bottom when a new item is added
-    LaunchedEffect(spaces.size) {
-        if (addItemJustOccurred) {
-            scrollState.scrollToItem(index = spaces.lastIndex)
-            addItemJustOccurred = false
-        }
-    }
-
-    // Library items
-    LazyColumn(
-        state = scrollState,
-        contentPadding = PaddingValues(
-            bottom = fabClearance,
-            top = 16.dp, start = 16.dp, end = 16.dp
-        ),
-        verticalArrangement = Arrangement.spacedBy(16.dp)
-    ) {
-
-        itemsIndexed(items = spaces) { index, spaceModel ->
-            LibraryArtSpaceItem(
-                spaceModel = spaceModel,
-                isNew = addItemJustOccurred && (spaces.lastIndex == index),
-                onDelete = onDelete,
-                onOpen = onOpen,
-            )
-        }
-
-    }
-
-    var openDialog by remember { mutableStateOf(false) }
-
-    // Fab, opens dialog to create item
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp)
-    ) {
-        AddItemFab(
-            onClick = { openDialog = true },
-            modifier = Modifier
-                .onGloballyPositioned { coordinates ->
-                    fabHeight = coordinates.size.height
-                }
-                .align(Alignment.BottomEnd)
-        )
-    }
-
-    // Dialog, creates item
-    if (openDialog)
-        CreateItemDialog(
-            onCreate = onCreate,
-            onCancel = { openDialog = false }
-        )
-
-}
-
-@Composable
-private fun LibraryArtSpaceItem(
-    spaceModel: SpaceModel,
-    isNew: Boolean,
-    onDelete: (SpaceModel) -> Unit,
-    onOpen: (SpaceModel) -> Unit
-) {
-
-    val artSpace = spaceModel.toArtSpace()
-    val cardShape = CardDefaults.shape
-
-    // Opacity animation for newly created items, out -> in
-    var targetAlpha by remember {
-        mutableStateOf(
-            when (isNew) {
-                true -> 0f; false -> 1f
-            }
-        )
-    }
-    val animatedAlpha: Float by animateFloatAsState(
-        targetValue = targetAlpha,
-        animationSpec = tween(
-            durationMillis = 1000,
-            easing = LinearOutSlowInEasing,
-        ),
-    )
-    // Begins the alpha float animation, 0f -> 1f
-    if (isNew) targetAlpha = 1f
-
-    Card(
-        shape = cardShape,
-        modifier = Modifier
-            .fillMaxWidth()
-            .alpha(alpha = animatedAlpha)
-    ) {
-
-        Column {
-            // Drawing image
-            BitmapImage(
-                pixelGrid = artSpace.state.colorDrawing,
-                contentDescription = "Localized description",
-                contentScale = ContentScale.Crop,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .clip(cardShape),
-            )
-            // Icon button row
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(all = 4.dp)
-            ) {
-                // Delete icon button
-                IconButton(
-                    onClick = { onDelete(spaceModel) }
-                ) {
-                    Icon(
-                        imageVector = Icons.Filled.Delete,
-                        contentDescription = "Localized description"
-                    )
-                }
-                // Dominant-hand icons
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth(),
-                    horizontalArrangement = Arrangement.End,
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    // Export icon button
-                    IconButton(onClick = { /* Handle click */ }) {
-                        Icon(
-                            imageVector = Icons.Filled.Save,
-                            contentDescription = "Localized description"
-                        )
-                    }
-                    // Share icon button
-                    IconButton(onClick = { /* Handle click */ }) {
-                        Icon(
-                            imageVector = Icons.Filled.Share,
-                            contentDescription = "Localized description"
-                        )
-                    }
-                    // Open icon button
-                    IconButton(
-                        onClick = { onOpen(spaceModel) }
-                    ) {
-                        Icon(
-                            imageVector = Icons.Filled.Draw,
-                            contentDescription = "Localized description"
-                        )
-                    }
-                }
-
-            } // End Box
-
-        } // End Column
-
-    } // End Card
-
-} // LibraryArtSpaceItem
-
-@Composable
-fun AddItemFab(
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    FloatingActionButton(
-        onClick = onClick,
-        modifier = modifier,
-    ) {
-        Icon(
-            imageVector = Icons.Filled.Add,
-            contentDescription = "Localized description"
-        )
-    }
-}
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-private fun CreateItemDialog(
+fun CreateItemDialog(
     onCreate: (width: Int, height: Int, paletteSize: Int) -> Unit,
     onCancel: () -> Unit
 ) {
@@ -478,20 +278,6 @@ private fun IntSlider(
     )
 }
 
-
-@Preview
-@Composable
-private fun LibraryScreenPreview() {
-    PreviewSurface {
-        val fauxModel = ArtSpace().toModel()
-        LibraryScreenContent(
-            spaces = listOf(fauxModel, fauxModel, fauxModel),
-            onCreate = { _, _, _ -> },
-            onDelete = {},
-            onOpen = {},
-        )
-    }
-}
 
 @Preview
 @Composable
